Move TypeInfo and FILE_TYPES to data_containers
